App/add checkout callback page

+++ services/product/src/handler/Product/delete.ts
@@ -1,7 +1,6 @@
-import { ProductId } from "@product/types/Product";
+import {ProductId, Recipe} from "@product/types/Product";
 import { Data } from "@product/types";
 import { log } from "@product/lib/log";
-import { PrismaClient } from "@prisma/client";
 import prisma from "@product/lib/prisma";
 import { ServerErrorResponse } from "@grpc/grpc-js";
 
@@ -11,6 +10,14 @@ export const DeleteProduct = async (
 ) => {
 	try {
 		const { id } = request;
+
+		const recipe = await prisma.recipe.findMany({where: {product_id: id}}) as Recipe[];
+	 	recipe.forEach(async (element) => {
+			await prisma.recipe.delete({where: {id: element.id}});
+		});
+
+		if((await prisma.product.findFirst({where: {id: id}})) == null)
+			throw(Error("Le produit n'existe  pas") as ServerErrorResponse)
 
 		await prisma.product.delete({ where : {id} });
